Step3 (#3496)
* docs : 기능 구현 목록 작성

* feat : 입력 받은 금액을 통해 구입금액을 생성하는 기능 구현

* feat : 구입 금액을 통해 구매한 로또 갯수를 출력하는 기능 구현

* feat : 로또 개수를 통해 로또 티켓을 발행하고 출력하는 기능 구현

* feat : 당첨 번호를 입력 받는 기능 구현

* feat : 발행 로또와 당첨 번호를 통해 당첨 통계를 구해 출력하는 기능 구현

* feat : 수익률을 생성해 출력하는 기능 구현

* refactor : 가독성 향상을 위한 메서드 분리

* refactor : 가독성 향상을 위한 네이밍 개선

* refactor : 로또 생성 객체 메소드 분리

* refactor : 로또 생성 객체의 부생성자 분리

* refactor : 리뷰 반영 리팩토링

* feat : 당첨 통계에 보너스 볼 합산 기능 추가 구현

* refactor : 사용하지 않는 클래스 삭제

* refactor : 생성자 접근제어자 수정

* refactor : 리뷰 반영 1차 리팩토링

+++ src/main/java/lotto/domain/PurchaseAmount.java
@@ -23,7 +23,7 @@ public class PurchaseAmount {
         return purchaseAmount % LOTTO_PRICE == 0;
     }
 
-    public double divideByPurchaseAmount(int prizeAmount) {
+    public double divideByPurchaseAmount(long prizeAmount) {
         return (double) prizeAmount / purchaseAmount;
     }
